Move update command to ConnectFour
Also move BrowserNotifier to shared libraries.
