+++ .env
@@ -6,6 +6,7 @@ APP_ENV='dev'
 APP_KERNEL_SECRET='ThisTokenIsNotSoSecretChangeIt'
 APP_WAIT_FOR='unix:///var/run/proxysql/proxysql.sock,mysql:3306,redis:6379,rabbitmq:5672,nchan:81'
 APP_RUN_MIGRATIONS='1'
+APP_NCHAN_BASE_URL='http://nchan:81'
 APP_RABBIT_MQ_DSN='amqp://guest:guest@rabbitmq:5672?heartbeat=60&prefetchCount=30'
 
 ############################
@@ -31,5 +32,4 @@ APP_IDENTITY_DOCTRINE_DBAL_URL='mysqli://root:password@localhost/identity?persis
 #  Web Interface Context   #
 ############################
 APP_WEB_INTERFACE_PREDIS_CLIENT_URL='redis://redis:6379?persistent=1'
-APP_WEB_INTERFACE_NCHAN_BASE_URL='http://nchan:81'
 APP_WEB_INTERFACE_PUBLISH_TO_BROWSER_SHARDS='1'
