+++ config/config.yml
@@ -108,6 +108,17 @@ services:
     gaming.scheduler:
         class: Gaming\Common\Scheduler\PcntlScheduler
 
+    gaming.browser-notifier:
+        class: Gaming\Common\BrowserNotifier\NchanBrowserNotifier
+        arguments:
+            - !service
+                class: Marein\Nchan\Nchan
+                arguments:
+                    - '%env(APP_NCHAN_BASE_URL)%'
+                    - !service
+                        class: Marein\Nchan\HttpAdapter\Psr18ClientAdapter
+                        arguments: ['@psr18.http_client', '@psr18.http_client', '@psr18.http_client']
+
     # Custom exception listener for the gaming domain.
     # Lower priority so that the profiler is respected.
     # todo: This can be removed as soon as https://github.com/marein/php-gaming-website/issues/34 is done.
