+++ config/connect-four/services/console.yml
@@ -12,3 +12,10 @@ services:
             -   name: console.command
                 command: connect-four:follow-event-store
                 description: 'Publish events to subscribers.'
+
+    Gaming\ConnectFour\Port\Adapter\Console\PublishRunningGamesCountToNchanCommand:
+        arguments: ['@connect-four.query-bus', '@gaming.browser-notifier']
+        tags:
+            -   name: console.command
+                command: connect-four:publish-running-games-count-to-nchan
+                description: 'Publish number of running games to the browser.'
