--- config/web-interface/services/integration.yml
@@ -1,12 +1,0 @@
-services:
-    web-interface.nchan:
-        class: Marein\Nchan\Nchan
-        arguments:
-            - '%env(APP_WEB_INTERFACE_NCHAN_BASE_URL)%'
-            - !service
-                class: Marein\Nchan\HttpAdapter\Psr18ClientAdapter
-                arguments: ['@psr18.http_client', '@psr18.http_client', '@psr18.http_client']
-
-    web-interface.browser-notifier:
-        class: Gaming\WebInterface\Infrastructure\NchanBrowserNotifier
-        arguments: ['@web-interface.nchan']
