+++ deploy/load-test/stack/app.env
@@ -15,7 +15,7 @@ APP_CONNECT_FOUR_PREDIS_CLIENT_URL=redis://connect-four-redis:6379?persistent=1
 APP_IDENTITY_DOCTRINE_DBAL_URL=mysqli://root:password@localhost/identity?persistent=1&unix_socket=/var/run/proxysql/proxysql.sock
 
 APP_WEB_INTERFACE_PREDIS_CLIENT_URL=redis://web-interface-redis:6379?persistent=1
-APP_WEB_INTERFACE_NCHAN_BASE_URL=http://nchan:81
+APP_NCHAN_BASE_URL=http://nchan:81
 APP_WEB_INTERFACE_PUBLISH_TO_BROWSER_SHARDS=5
 
 PHP_FPM_PM=static
