+++ deploy/load-test/stack/web-interface/app.yml
@@ -29,7 +29,7 @@ services:
                     - "node.labels.long-running==1"
     web-interface-publish-running-games-count-to-nchan:
         image: marein/php-gaming-website:php-fpm
-        command: bin/console web-interface:publish-running-games-count-to-nchan
+        command: bin/console connect-four:publish-running-games-count-to-nchan
         env_file: ../app.env
         volumes:
             - proxysql.sock:/var/run/proxysql
