+++ docker-compose.ci.yml
@@ -106,7 +106,7 @@ services:
         command: bin/console gaming:consume-messages all
     php-web-interface-publish-running-games-count-to-nchan:
         <<: *php-container
-        command: bin/console web-interface:publish-running-games-count-to-nchan
+        command: bin/console connect-four:publish-running-games-count-to-nchan
 
     ##############################
     #          CI helper         #
